requirements.txt
transformers
diffusers
torch
pydantic
boto3frontend